Migrate grid_builder and theme packages to src structure
