+++ packages/apptive_grid_theme/test/apptive_grid_theme_test.dart
@@ -1,5 +1,5 @@
 import 'package:alchemist/alchemist.dart';
-import 'package:apptive_grid_theme/apptive_grid_theme.dart';
+import 'package:apptive_grid_theme/src/apptive_grid_theme.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:golden_toolkit/golden_toolkit.dart' show loadAppFonts;
